+++ R/badge.R
@@ -36,6 +36,50 @@ use_depsy_badge <- function() {
   invisible(TRUE)
 }
 
+
+#' Create a life cycle badge
+#'
+#' * Experimental: very early days, a lot of churn in search of a good API.
+#'   Not on CRAN. Might not go anywhere. Use with care.
+#' * Maturing: API roughed out, but finer details likely to change. Will strive
+#'   to maintain backward compatibility, but need wider usage in order to get
+#'   more feedback.
+#' * Dormant: not currently working on it, but plan to come back to in the
+#'   future.
+#' * Stable: we're happy with the API, and the unlikely to be major changes.
+#'   Backward incompatible changes will only be made if absolutely critical, and
+#'   will be accompanied by a change in the major version.
+#' * Questioning: no long convinced this is a good approach, but don't yet
+#'   know what a better approach might be.
+#' * Retired: known better replacement available elsewhere. Will remain
+#'   available on CRAN.
+#' * Archived: development complete. Archived on CRAN and on GitHub.
+#'
+#' @param stage Stage of the lifecycle. See description above.
+#' @export
+use_lifecycle_badge <- function(stage) {
+
+  stage <- match.arg(tolower(stage), names(stages))
+  colour <- stages[[stage]]
+
+  url <- paste0(
+    "https://img.shields.io/badge/lifecycle-", stage, "-", colour, ".svg"
+  )
+
+  use_badge("lifecycle", url, url)
+}
+
+stages <- c(
+  experimental = "orange",
+  maturing = "blue",
+  stable = "brightgreen",
+  retired = "orange",
+  archived = "red",
+  dormant = "blue",
+  questioning = "blue"
+)
+
+
 #' Use a README badge
 #'
 #' @param badge_name Badge name. Used in error message and alt text
@@ -44,7 +88,7 @@ use_depsy_badge <- function() {
 #' @export
 use_badge <- function(badge_name, href, src) {
   if (has_badge(href)) {
-    return(FALSE)
+    return(invisible(FALSE))
   }
 
   img <- paste0("![", badge_name, "](", src, ")")