+++ R/browse.R
@@ -1,29 +1,38 @@
 #' Quickly browse to important package webpages
 #'
 #' @param package Name of package; leave as `NULL` to use current package
+#' @param number For GitHub issues and pull requests. Can be a number or
+#'   `"new"`.
 #' @export
 #' @examples
 #' browse_cran("MASS")
 browse_github <- function(package = NULL) {
-  view_url(github_home(package = package))
+  view_url(github_home(package))
 }
 
 #' @export
 #' @rdname browse_github
-browse_github_issues <- function(package = NULL) {
-  view_url(github_home(package = package), "/issues")
+browse_github_issues <- function(package = NULL, number = NULL) {
+  view_url(github_home(package), "issues", number)
 }
 
 #' @export
 #' @rdname browse_github
-browse_github_pulls <- function(package = NULL) {
-  view_url(github_home(package = package), "/pulls")
+browse_github_pulls <- function(package = NULL, number = NULL) {
+  view_url(github_home(package), "pulls", number)
+}
+
+#' @export
+#' @rdname browse_github
+browse_travis <- function(package = NULL) {
+  gh <- github_home(package)
+  view_url(sub("github.com", "travis-ci.org", gh))
 }
 
 #' @export
 #' @rdname browse_github
 browse_cran <- function(package = NULL) {
-  view_url(cran_home(package = package))
+  view_url(cran_home(package))
 }
 
 github_home <- function(package = NULL) {