+++ R/tidyverse.R
@@ -15,6 +15,9 @@
 #' * `use_tidy_eval()`: imports a standard set of helpers to facilitate
 #'   programming with the tidy eval toolkit.
 #'
+#' * `use_tidy_versions()`: pins all dependencies to require at least
+#'   the currently installed version.
+#'
 #' @md
 #' @name tidyverse
 NULL
@@ -66,6 +69,32 @@ use_tidy_description <- function() {
   desc::desc_reorder_fields(file = base_path)
 
   invisible(TRUE)
+}
+
+
+#' @export
+#' @rdname tidyverse
+#' @param overwrite By default (`FALSE`), only dependencies without version
+#'   specifications will be modified. Set to `TRUE` to modify all dependencies.
+use_tidy_versions <- function(overwrite = FALSE) {
+  deps <- desc::desc_get_deps(proj_get())
+
+  to_change <- deps$package != "R"
+  if (!overwrite) {
+    to_change <- to_change & deps$version == "*"
+  }
+
+  deps$version[to_change] <- vapply(deps$package[to_change], dep_version, character(1))
+  desc::desc_set_deps(deps, file = proj_get())
+
+  invisible(TRUE)
+}
+
+is_installed <- function(x) {
+  length(find.package(x, quiet = TRUE)) > 0
+}
+dep_version <- function(x) {
+  if (is_installed(x)) paste0(">= ", utils::packageVersion(x)) else "*"
 }